EmotionSong/requirements.txt
opencv-python
googlesearch-python
deepface
tensorflow